\name{StatIndex12R}
\alias{StatIndex12R}
\docType{data}
\title{
proto object for stat_index12
}
\description{
Create \code{stat_index12()}.
}
\usage{stat_index12 <- StatIndex12R$new }
\format{
  The format is:
proto object 
 $ calculate_groups:function (., data, scales, ...)   
 $ objname         : chr "index12" 
 $ default_geom    :function (.)   
 $ calculate       :function (., data, scales, ...)   
 $ required_aes    : chr [1:2] "x" "y" 
 parent: proto object 
 .. parent: proto object 
}
\details{
This object operates in the background to create \code{stat_index12()} and does not need to be called by the user.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{

}
\keyword{datasets}
