\name{StatIndex1R}
\alias{StatIndex1R}
\docType{data}
\title{
proto object for stat_index
}
\description{
Create \code{stat_index1()}.
}
\usage{stat_index1 <- StatIndex1R$new }
\format{
  The format is:
\code{proto object 
 $ calculate_groups:function (., data, scales, ...)   
 $ objname         : chr "index" 
 $ default_geom    :function (.)   
 $ calculate       :function (., data, scales, ...)   
 $ required_aes    : chr [1:2] "x" "y" 
 parent: proto object 
 .. parent: proto object }
}
\details{
This object operates in the background to create \code{stat_index1()} and does not need to be called by the user.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{

}
\keyword{datasets}
