\name{theme_nothing}
\alias{theme_nothing}
\title{
A blank theme for ggplot2 plots.
}
\description{
Removes axes, makes rectangles and lines blank, etc - so only elements that are explicitly redefined will be drawn.
}
\usage{
theme_nothing(base_size = 12, base_family = "Helvetica")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{base_size}{
Base font size.
}
  \item{base_family}{
Base font family.
}
}
\details{
}
\value{
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Nicked from Dragonfly code used for creating infographics
}

\note{
}

\seealso{

}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ aplot }


